shift to timed subshell execution to prevent environment wonkiness

+++ fio/microbench.sh
@@ -63,21 +63,27 @@ function onexit() {
 
 main () {
     trap onexit 0 # Havest/sigquit all subshells - forks() array
-    rm -rf "${resultsdir}" && mkdir -p "${resultsdir}"
-
-    commands=("/usr/bin/bash && time /usr/bin/git clone git@github.com:MicrosoftDocs/azure-docs.git $SCR")
+    mkdir -p "${resultsdir}"
 
 
 
     for directory in "${drive_dirs[@]}"; do
         cd "${directory}" || exit 1
         rm -rf "${directory:?}/*" || echo 'clear'
-        for comm in "${commands[@]}"; do
+
+        for filename in micro_tests/*.sh; do
+            [ -e "$filename" ] || continue
             scr="${directory}/scratch-temp"
             mkdir -p "${scr}"
-            echo "running   " "${comm}"
-            # Need to stop/start the watchers for each test
-            "export SCR=${scr} && env && $comm"
+            # Execute the command without timing to warm the cache
+            ${filename}
+            rm -rf "${scr}" && mkdir -p "${scr}"
+            # Run a loop of 5 iterations
+            for i in {0..5}; do
+                echo "${filename} iteration $i ============ "
+                /usr/bin/time -a -o "${filename}.time.out" -f "%E real,%U user,%S sys" "${filename} ${scr}"
+                rm -rf "${scr}" && mkdir -p "${scr}"
+            done
             rm -rf "${directory}/scratch-temp"
         done
     done
